requirements.txt
streamlit>=1.32.0
supabase>=2.3.0
google-genai>=1.0.0
pandas>=2.0.0
python-docx>=1.1.0
reportlab>=4.0.0
